refactor(...Repository): 인터페이스가 아닌 구현체만 사용하도록 변경
