+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,12 +1,27 @@
 package roomescape.repository;
 
 import java.util.List;
-import roomescape.domain.ReservationTime;
+import org.springframework.stereotype.Component;
+import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.Time;
 
-public interface ReservationTimeRepository {
-    void saveReservationTime(ReservationTime reservationTime);
+@Component
+public class ReservationTimeRepository {
+    private final ReservationTimeDao reservationTimeDao;
 
-    List<ReservationTime> findAllReservationTimes();
+    public ReservationTimeRepository(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
+    }
 
-    void deleteReservationTimeById(long reservationTimeId);
+    public void saveReservationTime(Time reservationTime) {
+        reservationTimeDao.save(reservationTime);
+    }
+
+    public List<Time> findAllReservationTimes() {
+        return reservationTimeDao.findAll();
+    }
+
+    public void deleteReservationTimeById(long reservationTimeId) {
+        reservationTimeDao.deleteById(reservationTimeId);
+    }
 }
